requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
coverage==7.9.2
distlib==0.4.0
pytest==8.4.0
pytest-cov==6.2.1
selenium==4.34.2
virtualenv==20.34.0

